vlm-b2b/requirements.txt
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.29.0
pandas>=2.0.0
python-dotenv>=1.0.0
stripe>=8.0.0
requests>=2.31.0
